rewrite to typescript
